// modules/dxp/apps/content-site-generator/content-site-generator-web/src/main/resources/META-INF/resources/js/types/GeneratedPage.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export interface GeneratedPage {
	icon: string;
	id: string;
	itemCount: number;
	languages: string[];
	templateLabel: string;
	title: string;
	url: string | null;
}